Kernel: Split the Ext2FileSystem.{cpp,h} files into smaller components
