+++ Kernel/Syscalls/mount.cpp
@@ -6,7 +6,7 @@
 
 #include <Kernel/FileSystem/Custody.h>
 #include <Kernel/FileSystem/DevPtsFS/FileSystem.h>
-#include <Kernel/FileSystem/Ext2FileSystem.h>
+#include <Kernel/FileSystem/Ext2FS/FileSystem.h>
 #include <Kernel/FileSystem/FATFS/FileSystem.h>
 #include <Kernel/FileSystem/ISO9660FS/FileSystem.h>
 #include <Kernel/FileSystem/Plan9FS/FileSystem.h>
